# doc/api.yaml
openapi: 3.0.3
info:
  title: WasaPhoto API
  description: |-
    This OpenAPI document describes the functionalities of a social media platform, WasaPhoto. The website has all the basic features such as banning/unbanning other users, liking/unliking photos, upload/delete photos and commenting/uncommenting on photos. 
    
    This API supports Cross-Origin Resource Sharing (CORS), allowing it to interact with other domains in this API for data reasons to increase the website's usability.
    
    This API was created in 2023/2024.
    
  version: "10"
paths:
  /session:
    post:
      tags: ["user"]
      summary: Log in the user
      description: |-
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user does exist, the user identifier is returned.

      operationId: doLogin
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              description: jlkjljòljò
              type: object
              properties:
                username:
                  type: string
                  example: Maria
                  pattern: 'ˆ.*?$'
                  minLength: 3
                  maxLength: 16
        required: true
      responses:
        '201':
          description: User log-in action successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  userID:
                    type: string
                    example: "abcdef012345"
  /user/{userID}/setusername/{username}:
    put:
      tags: ["user"]
      summary: Sets or updates the user's username 
      description: |-
        This endpoint allows the user to set their new username.
        If successful, it returns a message that the username has been updated. The username can not be set if it already exists.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: username
          required: true
          description: |-
            username for login
          schema:
            type: string
        - in: path
          name: userID
          required: true
          description: |-
            username for login
          schema:
            type: string
      operationId: setMyUserName
      requestBody:
        description: New username
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: Shireen2003
                  minLength: 5
                  maxLength: 15
        required: true
      responses:
        '200':
          description: The username has been updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Username"
        '400':
          description: Username already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "UserID not found"
...
  /photo/upload/{userID}:
    post:
      tags: ["photo"]
      summary: Upload a photo
      description: |-
        This endpoint allows a user to upload a photo.
      security:
        - bearerAuth: []     
      operationId: uploadPhoto
      parameters:
        - in: path
          name: userID
          required: true
          description: |-
            post photo on user's account by identifying them with their userID.
          schema:
            type: string
      requestBody:
        description: |-
          new photo file upload request. If successful, a new photoID is
          created. Each photo has a unique identifier which is the photoID.
        content:
          multipart/form-data:
              schema: 
                type: object
                properties:
                  file:
                    type: string
                    format: binary
                    
          
        required: true

      responses:
        '200':
          description: Photo uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
...
  /username/{username}/profile:
    get:
      tags: ["user"]
      summary: Get User Profile
      description: |-
        This endpoint allows you to retrieve your profile information by their userID.
      security:
        - bearerAuth: []
      operationId: getUserProfile
      parameters:
        - in: path
          name: username
          required: true
          description: Username of the person trying to retrieve the profile
          schema:
            type: string
            example: maria
            pattern: 'ˆ.*?$'
            minLength: 1
            maxLength: 32
 
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
                  
        
        '404':
          description: User profile not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "User profile not found"
  /user/{userID}/stream:
    get:
      tags: ["user"]
      summary: Get User's Stream
      description: |-
        This endpoint allows a user to retrieve their stream. A user's stream is a set of photos from other users that they follow. The stream of photos are arranged in reverse chronological order, which includes information about the date and time of each photo uploaded, the number of likes, and the comments on each post.
      security:
        - bearerAuth: []
      operationId: getMyStream
      parameters:
        - in: path
          name: userID
          required: true
          description: userID of the person's stream 
          schema:
            type: string
      responses:
        '200':
          description: User's stream acquired successfully
          content:
            image/png:
                schema:
                  $ref: "#/components/schemas/Stream"
                  x-sort:
                    - uploadDateTime: desc
                  
        '400':
          description: Stream information not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "profile stream unavailable"
...
  /photo/{photoID}/remove:
    delete:
      tags: ["photo"]
      summary: Delete a Photo
      description: |-
        This endpoint allows a user to delete a photo.
      security:
        - bearerAuth: []
      operationId: deletePhoto
      parameters:
        - in: path
          name: photoID
          required: true
          description: |-
            PhotoID is a unique identifier. This is needed to retrieve the
            specific photo to delete.
          schema:
            type: string
            example: "pic1"
      responses:
        '200':
          description: Photo is deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "photo removed successfully"
                
        '404':
          description: Photo not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "photoID not found"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |-
        App session token value. This value is sent by the server in the login
        and session refresh APIs.
  schemas:
    Username:
      type: object
      description: user's username
      properties:
        Username:
          type: string
          example: Maria
          pattern: 'ˆ.*?$'
          minLength: 3
          maxLength: 16
    Profile:
      type: object
      description: user profile
      properties:
      
        username:
          type: string
          example: Maria
          pattern: 'ˆ.*?$'
          minLength: 3
          maxLength: 16
          description: the username
          
        followCount:
          type: integer
          description: Amount of followers following the user.
          example: 24

        following: 
          type: integer
          description: Amount of followers the user is following
          example: 24
          
        photoCount:
          type: integer
          description: Amount of photos the user has posted
          example: 24
...
    Photo:
      type: object
      description: A photo
      properties:
        photoID: 
          type: integer
          description: the unique identifier for a photo
          minLength: 3
          maxLength: 500
          example: 1
        userID:
          type: string
          description: user identifier
          minLength: 10
          maxLength: 80
        photo:
          type: string
          format: binary
          minLength: 1
          maxLength: 100000
          
        like_count:
          type: integer
          description: The number of likes of the photo.
          example: 22
        comment_count:
          type: integer
          description: The number of comments of a photo.
          example: 22
        date:
          type: string
          format: date-time
          minLength: 20
          maxLength: 21
          description: the date and time when the photo is uploaded.
          pattern: '^.*?$'
          example: 2021-05-10T06:11:09Z
        username:
          type: string
          example: Maria
          pattern: 'ˆ.*?$'
          minLength: 3
          maxLength: 16
          description: the username
        isLiked:
          type: boolean
          example: true
          description: check whether the current user has liked the photo
    Stream:
      type : object
      description : stream of photos of the users the user follows.
      properties :
        photos: 
          description: list of the photos
          type: array
          minItems: 0
          maxItems: 10000000
          items:
            type: object
            properties:
              photoID: 
                description:  the photo identifier.
                type: string
                pattern: '^.*?$'
                example: "123-123-123"
                minLength: 3
                maxLength: 50
              userID:
                description:  user identifier of the photo posted by a user.
                type: string
                pattern: '^.*?$'
                example: "123-123-123" 
                minLength: 3
                maxLength: 50
                
              username:
                description: username.
                type: string
                pattern: '^.*?$'
                example: shireen123
                minLength: 3
                maxLength: 16
                
              likeCount:
                description: amount of likes on the a photo.
                type: integer
                example: 32
                minimum: 0
                maximum: 100000
                
              commentCount:
                description: amount of comments on a photo.
                type: integer
                example: 0
                minimum: 0
                maximum: 9939
                
  
              isLiked:
                description: check whether the current user liked the photo
                type: boolean
                example: false

              date:
                description: date of when the photo was uploaded
                type: string
                format: date-time
                example: "2021-06-06T02:01:01Z"
                minLength: 20
                maxLength: 40
                
              document: 
                description: the photo file. PNG or JPEG file
                type: string
                format: binary
                minLength: 3
                maxLength: 100
...